Chnged route

+++ app/services/users.py
@@ -30,9 +30,16 @@ class UsersService:
     async def update_user(self, user_id: int, user_data: UserUpdate):
         user_dict = user_data.model_dump()
         filter_by = {"id": user_id}
+        user = await self.users_repo.find_by_filter(filter_by)
+        if user is None:
+            raise HTTPException(status_code=404, detail=ERROR_USER_NOT_FOUND)
         return await self.users_repo.update_by_filter(filter_by, user_dict)
 
     async def delete_user(self, user_id: int):
+        filter_by = {"id": user_id}
+        user = await self.users_repo.find_by_filter(filter_by)
+        if user is None:
+            raise HTTPException(status_code=404, detail=ERROR_USER_NOT_FOUND)
         await self.users_repo.delete_by_id(user_id)
 
     async def find_user_by_email(self, email: str):
